Move reusable action code into its own module
This restores `project.py` to being a pure parsing module, but we'll
need to make some further changes to hook the reusable actions code back
into the execution path.
