+++ jobrunner/project.py
@@ -6,12 +6,6 @@ from pathlib import PurePosixPath, PureWindowsPath
 from types import SimpleNamespace
 
 from jobrunner import config
-from jobrunner.lib import git
-from jobrunner.lib.github_validators import (
-    GithubValidationError,
-    validate_branch_and_commit,
-    validate_repo_url,
-)
 from jobrunner.lib.yaml_utils import YAMLError, parse_yaml
 
 # The magic action name which means "run every action"
@@ -46,33 +40,16 @@ class InvalidPatternError(ProjectValidationError):
     pass
 
 
-class ReusableActionError(Exception):
-    """Represents a study developer-friendly reusable action error.
-
-    We raise this in preference to other, lower-level, errors because there's only so
-    much a study developer can do when there's an error with a reusable action.
-    """
-
-
 # Tiny dataclass to capture the specification of a project action
 @dataclasses.dataclass
 class ActionSpecifiction:
     run: str
     needs: list
     outputs: dict
-    repo_url: str
-    commit: str
-
-
-@dataclasses.dataclass
-class ReusableAction:
-    repo_url: str
-    commit: str
-    action_file: bytes
 
 
 def parse_and_validate_project_file(project_file):
-    """Parse and validate the project file, handling reusable actions.
+    """Parse and validate the project file.
 
     Args:
         project_file: The contents of the project file as an immutable array of bytes.
@@ -87,136 +64,7 @@ def parse_and_validate_project_file(project_file):
         project = parse_yaml(project_file, name="project.yaml")
     except YAMLError as e:
         raise ProjectValidationError(*e.args)
-    actions = project["actions"]
-    for action_id, action in actions.items():
-        actions[action_id] = handle_reusable_action(action_id, action)
-    project = validate_project_and_set_defaults(project)
-    return project
-
-
-def handle_reusable_action(action_id, action):
-    """If `action` is reusable, then handle it. If not, then return it unchanged.
-
-    Args:
-        action_id: The action's ID as a string. This is the action's key in
-            project.yaml. It is used to raise errors with more informative messages.
-        action: The action's representation as a dict. This is the action's value in
-            project.yaml.
-
-    Returns:
-        The action's representation as a dict. If `action` resolves to a reusable
-        action, then it is rewritten to point to the reusable action and a copy is
-        returned. If not, then `action` is returned unchanged.
-
-    Raises:
-        ReusableActionError: An error occurred when accessing the reusable action.
-    """
-    run_args = shlex.split(action["run"])
-    image, tag = run_args[0].split(":")
-
-    if image in config.ALLOWED_IMAGES:
-        # This isn't a reusable action.
-        return action
-
-    reusable_action = fetch_reusable_action(action_id, image, tag)
-    new_action = apply_reusable_action(action_id, action, reusable_action)
-    return new_action
-
-
-def fetch_reusable_action(action_id, image, tag):
-    """
-    Fetch all metadata from git needed to apply a reusable action
-
-    Args:
-        action_id: The action's ID as a string. This is the action's key in
-            project.yaml. It is used to raise errors with more informative messages.
-        image: The name of the reusable action
-        tag: The specified version of the reusable action
-
-    Returns:
-        ReusableAction object, wrapping the repo_url, commit and the contents
-        of the `action.yaml` file
-
-    Raises:
-        ReusableActionError: An error occurred when accessing the reusable action.
-    """
-    repo_url = f"{config.ACTIONS_GITHUB_ORG_URL}/{image}"
-    try:
-        validate_repo_url(repo_url, [config.ACTIONS_GITHUB_ORG])
-    except GithubValidationError as e:
-        raise ReusableActionError(*e.args)  # This keeps the function signature clean
-
-    try:
-        # If there's a problem, then it relates to the repository. Maybe the study
-        # developer made an error; maybe the reusable action developer made an error.
-        commit = git.get_sha_from_remote_ref(repo_url, tag)
-    except git.GitError:
-        raise ReusableActionError(
-            f"Cannot resolve '{action_id}' to a repository at '{repo_url}'"
-        )
-
-    try:
-        validate_branch_and_commit(repo_url, commit, "main")
-    except GithubValidationError as e:
-        raise ReusableActionError(*e.args)
-
-    try:
-        # If there's a problem, then it relates to the reusable action. The study
-        # developer didn't make an error; the reusable action developer did.
-        action_file = git.read_file_from_repo(repo_url, commit, "action.yaml")
-    except git.GitError:
-        raise ReusableActionError(
-            f"There is a problem with the reusable action required by '{action_id}'"
-        )
-
-    return ReusableAction(repo_url=repo_url, commit=commit, action_file=action_file)
-
-
-def apply_reusable_action(action_id, action, reusable_action):
-    """
-    Rewrite an `action` dict to run the code specifed by the supplied
-    `ReusableAction` instance.
-
-    Args:
-        action_id: The action's ID as a string. This is the action's key in
-            project.yaml. It is used to raise errors with more informative messages.
-        action: The action's representation as a dict. This is the action's value in
-            project.yaml.
-        reusable_action: A ReusableAction instance
-
-    Returns:
-        The modified action's representation as a dict.
-
-    Raises:
-        ReusableActionError: An error occurred when accessing the reusable action.
-    """
-    try:
-        # If there's a problem, then it relates to the reusable action. The study
-        # developer didn't make an error; the reusable action developer did.
-        action_config = parse_yaml(reusable_action.action_file, name="action.yaml")
-        assert "run" in action_config
-        action_run_args = shlex.split(action_config["run"])
-        action_image, action_tag = action_run_args[0].split(":")
-        if action_image not in config.ALLOWED_IMAGES:
-            raise ProjectValidationError(f"Unrecognised runtime: {action_image}")
-        if is_generate_cohort_command(action_run_args):
-            raise ProjectValidationError(
-                "Re-usable actions cannot invoke cohortextractor"
-            )
-    except (YAMLError, AssertionError, ProjectValidationError):
-        raise ReusableActionError(
-            f"There is a problem with the reusable action required by '{action_id}'"
-        )
-
-    # ["action:tag", "arg", ...] -> ["runtime:tag binary entrypoint", "arg", ...]
-    run_args = shlex.split(action["run"])
-    run_args[0] = action_config["run"]
-
-    new_action = action.copy()
-    new_action["run"] = " ".join(run_args)
-    new_action["repo_url"] = reusable_action.repo_url
-    new_action["commit"] = reusable_action.commit
-    return new_action
+    return validate_project_and_set_defaults(project)
 
 
 # Copied almost verbatim from the original job-runner
@@ -394,10 +242,6 @@ def get_action_specification(project, action_id):
         run=run_command,
         needs=action_spec.get("needs", []),
         outputs=action_spec["outputs"],
-        # If action_spec (a dict) represents a reusable action, then it will have the
-        # following keys. If not, then it won't.
-        repo_url=action_spec.get("repo_url"),
-        commit=action_spec.get("commit"),
     )
