+++ tests/test_project.py
@@ -1,150 +1,15 @@
 import argparse
 import shlex
-from unittest import mock
 
 import pytest
 
 from jobrunner import project
-from jobrunner.lib import git
-from jobrunner.lib.yaml_utils import YAMLError
 from jobrunner.project import (
     InvalidPatternError,
     ProjectValidationError,
-    ReusableAction,
     assert_valid_glob_pattern,
     parse_and_validate_project_file,
 )
-
-
-@mock.patch.multiple(
-    "jobrunner.lib.git",
-    get_sha_from_remote_ref=mock.DEFAULT,
-    read_file_from_repo=mock.DEFAULT,
-)
-class TestHandleReusableAction:
-    def test_when_not_a_reusable_action(self, **kwargs):
-        # Happy path 1
-        action_in = {"run": "python:latest python analysis/my_action.py"}
-        action_out = project.handle_reusable_action("my_action", action_in)
-        assert action_in is action_out
-        kwargs["get_sha_from_remote_ref"].assert_not_called()
-        kwargs["read_file_from_repo"].assert_not_called()
-
-    @mock.patch(
-        "jobrunner.project.parse_yaml",
-        return_value={"run": "python:latest python reusable_action/main.py"},
-    )
-    def test_when_a_reusable_action_with_options(self, *args, **kwargs):
-        # Happy path 2
-        action_in = {"run": "reusable-action:latest --output-format=png"}
-        action_out = project.handle_reusable_action("my_action", action_in)
-        assert action_in is not action_out
-        assert (
-            action_out["run"]
-            == "python:latest python reusable_action/main.py --output-format=png"
-        )
-
-    @mock.patch(
-        "jobrunner.project.parse_yaml",
-        return_value={"run": "python:latest python reusable_action/main.py"},
-    )
-    def test_when_a_reusable_action_with_arguments(self, *args, **kwargs):
-        # Happy path 3
-        action_in = {"run": "reusable-action:latest output/input.csv"}
-        action_out = project.handle_reusable_action("my_action", action_in)
-        assert action_in is not action_out
-        assert (
-            action_out["run"]
-            == "python:latest python reusable_action/main.py output/input.csv"
-        )
-
-    @mock.patch(
-        "jobrunner.project.parse_yaml",
-        return_value={"run": "python:latest python reusable_action/main.py"},
-    )
-    def test_when_a_reusable_action_with_options_and_arguments(self, *args, **kwargs):
-        # Happy path 4
-        # We'll use Click's terminology.
-        # * Options are optional
-        # * Arguments are optional within reason, but are more restricted than options
-        # For more information, see:
-        # https://click.palletsprojects.com/en/8.0.x/parameters/
-        action_in = {
-            "run": "reusable-action:latest --output-format=png output/input.csv"
-        }
-        action_out = project.handle_reusable_action("my_action", action_in)
-        assert action_in is not action_out
-        assert (
-            action_out["run"]
-            == "python:latest python reusable_action/main.py --output-format=png output/input.csv"
-        )
-
-    def test_with_bad_run_command(self, **kwargs):
-        # We don't need to check the scheme, netloc, or org because we add those.
-        with pytest.raises(project.ReusableActionError):
-            project.handle_reusable_action(
-                "my_action",
-                {"run": "../my-bad-org/reusable-action:latest"},
-            )
-
-    def test_with_bad_remote_ref(self, **kwargs):
-        kwargs["get_sha_from_remote_ref"].side_effect = git.GitError
-        with pytest.raises(project.ReusableActionError):
-            project.handle_reusable_action(
-                "my_action", {"run": "reusable-action:latest"}
-            )
-
-    @mock.patch(
-        "jobrunner.project.validate_branch_and_commit",
-        side_effect=project.GithubValidationError,
-    )
-    def test_with_bad_commit(self, *args, **kwargs):
-        with pytest.raises(project.ReusableActionError):
-            project.handle_reusable_action(
-                "my_action",
-                {"run": "reusable-action:latest"},
-            )
-
-    def test_with_bad_file(self, **kwargs):
-        kwargs["read_file_from_repo"].side_effect = git.GitError
-        with pytest.raises(project.ReusableActionError):
-            project.handle_reusable_action(
-                "my_action", {"run": "reusable-action:latest"}
-            )
-
-    @mock.patch(
-        "jobrunner.project.parse_yaml",
-        side_effect=YAMLError,
-    )
-    def test_with_bad_yaml(self, *args, **kwargs):
-        with pytest.raises(project.ReusableActionError):
-            project.handle_reusable_action(
-                "my_action", {"run": "reusable-action:latest"}
-            )
-
-    @mock.patch("jobrunner.project.parse_yaml", return_value={})
-    def test_with_bad_action_config(self, *args, **kwargs):
-        with pytest.raises(project.ReusableActionError):
-            project.handle_reusable_action(
-                "my_action", {"run": "reusable-action:latest"}
-            )
-
-    def test_reusable_action_with_invalid_runtime(self, *args, **kwargs):
-        action_id = "my_action"
-        action = {"run": "foo:v1"}
-        reusable_action_1 = ReusableAction(
-            repo_url="foo", commit="bar", action_file=b"run: notanaction:v1"
-        )
-        with pytest.raises(project.ReusableActionError):
-            project.apply_reusable_action(action_id, action, reusable_action_1)
-        # This is a valid runtime, but it's not allowed in re-usable actions
-        reusable_action_2 = ReusableAction(
-            repo_url="foo",
-            commit="bar",
-            action_file=b"run: cohortextractor:v1 generate_cohort",
-        )
-        with pytest.raises(project.ReusableActionError):
-            project.apply_reusable_action(action_id, action, reusable_action_2)
 
 
 class TestParseAndValidateProjectFile:
@@ -163,32 +28,6 @@ class TestParseAndValidateProjectFile:
         project = parse_and_validate_project_file(project_file)
         obs_run = project["actions"]["my_action"]["run"]
         exp_run = "python:latest python analysis/my_action.py"
-        assert obs_run == exp_run
-
-    @mock.patch.multiple(
-        "jobrunner.lib.git",
-        get_sha_from_remote_ref=mock.DEFAULT,
-        read_file_from_repo=mock.DEFAULT,
-    )
-    def test_with_reusable_action(self, **kwargs):
-        project_file = """
-        version: '3.0'
-        expectations:
-            population_size: 1000
-        actions:
-            my_action:
-                run: reusable-action:latest --output-format=png
-                outputs:
-                    moderately_sensitive:
-                        my_figure: output/my_figure.png
-        """
-        action_file = """
-        run: python:latest python reusable_action/main.py
-        """
-        kwargs["read_file_from_repo"].return_value = action_file
-        project = parse_and_validate_project_file(project_file)
-        obs_run = project["actions"]["my_action"]["run"]
-        exp_run = "python:latest python reusable_action/main.py --output-format=png"
         assert obs_run == exp_run
 
     def test_with_duplicate_keys(self):
